refactor: separate pagination logic to custom hook
